Rename plugin files from chram-base to wp-normalize
I transferred this repo from Bitbucket to Github and in the meanwhile renamed it
to a bit more general name.
The idea of this plugin is to be a kind of CSS reset / normalize for WordPress.
It strips some parts I never use or don't want my users to see/use and adds some
things, like my phone number on the bottom of every dashboard page and extends
user rights for the editor role.